Refactor sync parallel-any strategy handling
