+++ projects/04-llm-adapter/tests/test_parallel_any_executor.py
@@ -40,6 +40,17 @@ except ImportError:  # pragma: no cover - RunnerMode 未導入環境向け
 PARALLEL_ANY_VALUE = RunnerMode.PARALLEL_ANY.value.replace("-", "_")
 
 
+from src.llm_adapter.errors import AllFailedError
+from src.llm_adapter.parallel_exec import run_parallel_all_sync, run_parallel_any_sync
+from src.llm_adapter.provider_spi import ProviderRequest, ProviderResponse, ProviderSPI
+from src.llm_adapter.runner_config import RunnerConfig as SyncRunnerConfig
+from src.llm_adapter.runner_config import RunnerMode as SyncRunnerMode
+from src.llm_adapter.runner_sync import Runner as SyncRunner
+from src.llm_adapter.runner_sync_modes import get_sync_strategy, SyncRunContext
+from src.llm_adapter.runner_sync_parallel_any import ParallelAnyStrategy
+from src.llm_adapter.utils import content_hash
+
+
 class FakeParallelExecutionError(RuntimeError):
     def __init__(self, message: str) -> None:
         super().__init__(message)
@@ -307,3 +318,60 @@ def test_parallel_any_mode_accepts_hyphen_compatibility(tmp_path: Path) -> None:
     index, result = batch[0]
     assert index == 0
     assert result.raw_output == "ok"
+
+
+def test_get_sync_strategy_parallel_any_propagates_all_failed(tmp_path: Path) -> None:
+    class _StubProvider(ProviderSPI):
+        def __init__(self, name: str) -> None:
+            self._name = name
+
+        def name(self) -> str:
+            return self._name
+
+        def capabilities(self) -> set[str]:
+            return set()
+
+        def invoke(self, request: ProviderRequest) -> ProviderResponse:  # pragma: no cover - not invoked
+            raise AssertionError("provider should not be invoked")
+
+    provider = _StubProvider("stub")
+    metrics_path = tmp_path / "metrics.jsonl"
+    config = SyncRunnerConfig(
+        mode=SyncRunnerMode.PARALLEL_ANY,
+        max_attempts=0,
+        metrics_path=str(metrics_path),
+    )
+    runner = SyncRunner([provider], config=config)
+    request = ProviderRequest(model="test", prompt="hello")
+    request_fingerprint = content_hash(
+        "runner",
+        request.prompt_text,
+        request.options,
+        request.max_tokens,
+    )
+    metadata = {
+        "run_id": request_fingerprint,
+        "mode": config.mode.value,
+        "providers": [provider.name()],
+        "shadow_used": False,
+        "shadow_provider_id": None,
+    }
+    context = SyncRunContext(
+        runner=runner,
+        request=request,
+        event_logger=None,
+        metadata=metadata,
+        run_started=0.0,
+        request_fingerprint=request_fingerprint,
+        shadow=None,
+        shadow_used=False,
+        metrics_path=str(metrics_path),
+        run_parallel_all=run_parallel_all_sync,
+        run_parallel_any=run_parallel_any_sync,
+    )
+
+    strategy = get_sync_strategy(config.mode)
+    assert isinstance(strategy, ParallelAnyStrategy)
+
+    with pytest.raises(AllFailedError):
+        strategy.execute(context)
